Ali_Project/KhmerOCR-main/requirements.txt
# Image Processing
opencv-python==4.8.0.76
Pillow==10.0.0
pdf2image==1.16.3

# OCR
pytesseract==0.3.10

# Machine Learning
joblib==1.3.1
numpy==1.24.3

# Document Processing
python-docx==0.8.11

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# System Utilities
psutil==5.9.5 